chore: add react-query
