+++ apps/web/package.json
@@ -6,6 +6,7 @@
     "dev": "next dev",
     "dev:mock": "concurrently --kill-others  \"NEXT_PUBLIC_API_MOCKING=enable next dev\" \"npx tsx watch ./src/mocks/http.ts\"",
     "generate:svg-component": "npx @svgr/cli --ignore-existing --jsx-runtime=automatic --out-dir src/components/common/icons --typescript -- src/assets/icons",
+    "code-gen": "pnpm graphql-codegen --config codegen.ts",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
@@ -18,6 +19,7 @@
     "@apollo/client": "^3.9.5",
     "@apollo/experimental-nextjs-app-support": "^0.11.2",
     "@sentry/nextjs": "^7.105.0",
+    "@tanstack/react-query": "^5.51.23",
     "@vercel/speed-insights": "^1.0.11",
     "class-variance-authority": "^0.7.0",
     "classnames": "^2.5.1",
@@ -63,6 +65,7 @@
     "@storybook/react": "^7.6.17",
     "@storybook/test": "^7.6.17",
     "@svgr/cli": "^8.1.0",
+    "@tanstack/react-query-devtools": "^5.51.23",
     "@types/express": "^4.17.21",
     "@types/gtag.js": "^0.0.19",
     "@types/jsonwebtoken": "^9.0.5",
